rename: eventTypes.ts → types.ts
